Caret position updated when right-clicking, typewriter next improved
Right-clicking on text now places the caret there and shows a context menu which matches the tags. Typewriter style down movement now also preserves the caret position. Functionality for obtaining the caret position in pixels written. The functionality for obtaining a right click position, however, does not (yet) work if the contents of the edit box have been scrolled.