"typewriter movement" improved
The caret has to be in the dialog in order to move and hence it's now placed there upon opening it. Many other improvements to navigation with the dialog. The new "move image instead of dialog behaviour" made step functions redundant.

Using "i" as a global variable is very problematic and one local i -> c substitution has been made but the global variable is the one whose name needs to change.