Edit dialog kept open when navigating
Issue no. 62 in -mc (!) and #82 in -edit at least partially resolved.
Among other things pathWithoutPage was not set correctly but should now
work. Additionally the dialog is repositioned slightly better if a user
resizes the window (but not perfectly in the case of "too small"
windows, i.e. when the dialog no longer fits).